fix(search): Disable refresh when searched
Disabled the refresh functionality (via setInterval) when the user
fills in the search bar. The refresh comes back when the search
bar is empty.
Moved the refreshTime in a variable so that it may be easier to
put in a config.
Moved the setInterval call in a method.
Added comments