feat(errors): Add outside territory error option
Added a new optionnal error that checks if an imported observation
is outside a territory.
The territory may be specified in conf_gn_module.toml and if not
provided, there will not be any check and so no errors.
Updated the data.sql and added a migration script to add a new
line in the t_user_errors table.
Moved some function to lighten the code