fix(import): Fix bug for field model import
When the user imports a json, a new model is created and the
form is filled. BUT: the select widget must be refresh to select
the new model and since it is done concurrently, the json load
was not done. Fix this by reading the json at the end. It will
be more difficult with nomenclatures...