fix(import): Fix nomenc field model import
Fixed the nomenclature field import via json. Find that there is a
subscribe in the content-mapping.service.ts which is not advised.
So replaced it with a map and call subscribe in the component
directly. Cleaner and there are no race between the form fields
actualisation and their cleaning. First clean the fill the fields
with the json file data